Pull out RunOptions and validation to own file
